update to mollify Elixir 1.2 compiler warnings
Prior to this commit, when compiled with Elixir 1.2, the compiler would
generate various warnings, mostly about ambiguous parameters near `|>`.
This commit adds parens to be explicit about additional parameters when
used with `|>`. This commit also adds a few other minor changes in
response to a handful of other minor compiler warnings.